expression,util/stringutil: move LIKE pattern match to stringutil package
